fix: cast PG AVG/PERCENTILE_CONT result to float8 to match Rust f64

+++ src/db/pg_backend.rs
@@ -2851,7 +2851,7 @@ impl DbBackend for PgBackend {
 
     async fn channel_usage_24h(&self) -> Result<Vec<(String, String, u64, u64, f64, f64)>, DbError> {
         let rows = sqlx::query_as::<_, (String, String, i64, i64, f64, f64)>(
-            "SELECT channel_id, model, COUNT(*)::bigint, SUM(CASE WHEN success THEN 1 ELSE 0 END)::bigint, COALESCE(AVG(latency_ms), 0), COALESCE(PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY latency_ms), 0)
+            "SELECT channel_id, model, COUNT(*)::bigint, SUM(CASE WHEN success THEN 1 ELSE 0 END)::bigint, COALESCE(AVG(latency_ms)::float8, 0), COALESCE(PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY latency_ms)::float8, 0)
              FROM usage_logs
              WHERE timestamp::timestamptz >= NOW() - INTERVAL '1 day'
              GROUP BY channel_id, model ORDER BY COUNT(*) DESC"
